Fixes some linting import/export issues
