+++ x-pack/plugins/observability/public/data_handler.ts
@@ -4,28 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { ObservabilityFetchDataResponse, FetchDataResponse } from './typings/fetch_data_response';
+import { DataHandler } from './typings/fetch_overview_data';
 import { ObservabilityApp } from '../typings/common';
-
-export interface FetchDataParams {
-  // The start timestamp in milliseconds of the queried time interval
-  startTime: string;
-  // The end timestamp in milliseconds of the queried time interval
-  endTime: string;
-  // The aggregation bucket size in milliseconds if applicable to the data source
-  bucketSize: string;
-}
-
-export type FetchData<T extends FetchDataResponse = FetchDataResponse> = (
-  fetchDataParams: FetchDataParams
-) => Promise<T>;
-
-export type HasData = () => Promise<boolean>;
-
-interface DataHandler<T extends ObservabilityApp = ObservabilityApp> {
-  fetchData: FetchData<ObservabilityFetchDataResponse[T]>;
-  hasData: HasData;
-}
 
 const dataHandlers: Partial<Record<ObservabilityApp, DataHandler>> = {};
