+++ x-pack/plugins/observability/public/typings/index.ts
@@ -6,4 +6,4 @@
 
 export * from './eui_draggable';
 export * from './eui_styled_components';
-export * from './fetch_data_response';
+export * from './fetch_overview_data';
